mount: Add OSX ACL translation layer
This commit makes LizardFS client on osx parse com.apple.system.Security
access control entries, translate them to RichACL and send to master.
Asking for com.apple.system.Security xattr on osx will result in
returning RichACL-s translated back.

+++ src/mount/fuse/mfs_fuse.cc
@@ -66,13 +66,13 @@ void getBsdGroups(LizardClient::Context &ctx) {
 #if defined(__APPLE__)
 		ctx.gids.resize(kinfo.kp_eproc.e_ucred.cr_ngroups + 1);
 		ctx.gids[0] = ctx.gid;
-		for (size_t i = 0; i < kinfo.kp_eproc.e_ucred.cr_ngroups; ++i) {
+		for (int i = 0; i < kinfo.kp_eproc.e_ucred.cr_ngroups; ++i) {
 			ctx.gids[i + 1] = kinfo.kp_eproc.e_ucred.cr_groups[i];
 		}
 #else
 		ctx.gids.resize(kinfo.ki_ngroups + 1);
 		ctx.gids[0] = ctx.gid;
-		for (size_t i = 0; i < kinfo.ki_ngroups; ++i) {
+		for (int i = 0; i < kinfo.ki_ngroups; ++i) {
 			ctx.gids[i + 1] = kinfo.ki_groups[i];
 		}
 #endif
